+++ src/solver/optimisation/constraints/STReleaseMaxReserve.cpp
@@ -27,9 +27,8 @@ void STReleaseMaxReserve::add(int pays, int reserve, int cluster, int pdt)
             1.0)
           .lessThan();
         data.CorrespondanceCntNativesCntOptim[pdt]
-          .reservesIndices.value()
-          .STStorageClusterMaxReleaseParticipation[reserveParticipation
-                                                     .globalIndexClusterParticipation]
+          .reservesIndices->STStorageClusterMaxReleaseParticipation
+            [reserveParticipation.globalIndexClusterParticipation]
           = builder.data.nombreDeContraintes;
 
         ConstraintNamer namer(builder.data.NomDesContraintes);